web: move WebGL shaders into web-common.
