+++ build.zig
@@ -1255,6 +1255,8 @@ fn addWebExample(
     compile_step.dependOn(&b.addInstallFileWithDir(output, install_dir, "index.html").step);
     const web_js = b.path("src/backends/web.js");
     compile_step.dependOn(&b.addInstallFileWithDir(web_js, install_dir, "web.js").step);
+    const web_common = b.path("src/backends/web-common.js");
+    compile_step.dependOn(&b.addInstallFileWithDir(web_common, install_dir, "web-common.js").step);
     b.addNamedLazyPath("web.js", web_js);
     compile_step.dependOn(&install_wasm.step);
     compile_step.dependOn(&install_noto.step);
@@ -1302,7 +1304,7 @@ fn addWebStandaloneExample(
     const compile_step = b.step(name, "Compile " ++ name ++ " (Web Worker standalone)");
     compile_step.dependOn(&b.addInstallFileWithDir(b.path("src/backends/index-standalone.html"), install_dir, "index.html").step);
     compile_step.dependOn(&b.addInstallFileWithDir(b.path("src/backends/web-standalone.js"), install_dir, "web-standalone.js").step);
-    compile_step.dependOn(&b.addInstallFileWithDir(b.path("src/backends/web-standalone-common.js"), install_dir, "web-standalone-common.js").step);
+    compile_step.dependOn(&b.addInstallFileWithDir(b.path("src/backends/web-common.js"), install_dir, "web-common.js").step);
     compile_step.dependOn(&b.addInstallFileWithDir(b.path("src/backends/web.js"), install_dir, "web.js").step);
     compile_step.dependOn(&b.addInstallFileWithDir(b.path("src/backends/web-worker.js"), install_dir, "web-worker.js").step);
     compile_step.dependOn(&install_wasm.step);
