+++ src/backends/web.js
@@ -53,85 +53,12 @@ async function dvui_open_file_picker(accept, multiple) {
     });
 }
 
-const vertexShaderSource_webgl = `
-    precision mediump float;
-
-    attribute vec4 aVertexPosition;
-    attribute vec4 aVertexColor;
-    attribute vec2 aTextureCoord;
-
-    uniform mat4 uMatrix;
-
-    varying vec4 vColor;
-    varying vec2 vTextureCoord;
-
-    void main() {
-      gl_Position = uMatrix * aVertexPosition;
-      vColor = aVertexColor / 255.0;  // normalize u8 colors to 0-1
-      vTextureCoord = aTextureCoord;
-    }
-`;
-
-const vertexShaderSource_webgl2 = `# version 300 es
-
-    precision mediump float;
-
-    in vec4 aVertexPosition;
-    in vec4 aVertexColor;
-    in vec2 aTextureCoord;
-
-    uniform mat4 uMatrix;
-
-    out vec4 vColor;
-    out vec2 vTextureCoord;
-
-    void main() {
-      gl_Position = uMatrix * aVertexPosition;
-      vColor = aVertexColor / 255.0;  // normalize u8 colors to 0-1
-      vTextureCoord = aTextureCoord;
-    }
-`;
-
-const fragmentShaderSource_webgl = `
-    precision mediump float;
-
-    varying vec4 vColor;
-    varying vec2 vTextureCoord;
-
-    uniform sampler2D uSampler;
-    uniform bool useTex;
-
-    void main() {
-        if (useTex) {
-            gl_FragColor = texture2D(uSampler, vTextureCoord) * vColor;
-        }
-        else {
-            gl_FragColor = vColor;
-        }
-    }
-`;
-
-const fragmentShaderSource_webgl2 = `# version 300 es
-
-    precision mediump float;
-
-    in vec4 vColor;
-    in vec2 vTextureCoord;
-
-    uniform sampler2D uSampler;
-    uniform bool useTex;
-
-    out vec4 fragColor;
-
-    void main() {
-        if (useTex) {
-            fragColor = texture(uSampler, vTextureCoord) * vColor;
-        }
-        else {
-            fragColor = vColor;
-        }
-    }
-`;
+import {
+    vertexShaderSource_webgl2,
+    vertexShaderSource_webgl,
+    fragmentShaderSource_webgl2,
+    fragmentShaderSource_webgl
+} from "./web-common.js";
 
 /**
  * @param {string | HTMLCanvasElement} canvas - A canvas element or string id of one
